+++ inst/ambiorix.js
@@ -8,6 +8,7 @@ if (window.location.protocol == "https:")
 // get websocket
 let ambiorixSocket = new WebSocket(protocol + window.location.host);
 
+// ambiorix Websocket
 class Ambiorix {
   constructor(){
     this._handlers = new Map()
@@ -23,7 +24,7 @@ class Ambiorix {
   }
   start(){
     var that = this;
-    ambiorixSocket.onmessage = function(msg){
+    ambiorixSocket.onmessage = (msg) => {
       let msgParsed = JSON.parse(msg.data);
 
       if(!msgParsed.isAmbiorix)
@@ -38,4 +39,21 @@ class Ambiorix {
   receive(name, fun){
     this._handlers.set(name, fun)
   }
+}
+
+// helper function to parseCookies
+// parseCookie(document.cookie);
+const parseCookie = str => {
+  if(str == "")
+    return {};
+
+  str
+    .split(';')
+    .map(v => v.split('='))
+    .reduce(
+      (acc, v) => {
+        acc[decodeURIComponent(v[0].trim())] = decodeURIComponent(v[1].trim());
+        return acc;
+      }, 
+    {});
 }